requirements.txt
2captcha-python==1.5.1
asgiref==3.8.1
attrs==25.3.0
beautifulsoup4==4.13.3
blinker==1.9.0
Brotli==1.1.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.1
contourpy==1.3.0
cryptography==44.0.2
cycler==0.12.1
Django==4.2.20
django-role-permissions==3.2.0
exceptiongroup==1.2.2
fonttools==4.57.0
h11==0.14.0
h2==4.2.0
hpack==4.1.0
hyperframe==6.1.0
idna==3.10
importlib_resources==6.5.2
joblib==1.4.2
kaitaistruct==0.10
kiwisolver==1.4.7
lxml==5.3.2
matplotlib==3.9.4
numpy==2.0.2
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.3
pillow==11.2.1
psycopg2-binary==2.9.10
pyasn1==0.6.1
pycparser==2.22
pyOpenSSL==25.0.0
pyparsing==3.2.3
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.1.0
pytz==2025.2
requests==2.32.3
scikit-learn==1.6.1
scipy==1.13.1
selenium==4.31.0
selenium-stealth==1.0.6
selenium-wire==5.1.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
sqlparse==0.5.3
threadpoolctl==3.6.0
trio==0.29.0
trio-websocket==0.12.2
typing_extensions==4.13.1
tzdata==2024.1
urllib3==1.26.18
webdriver-manager==4.0.2
websocket-client==1.8.0
wsproto==1.2.0
zipp==3.21.0
zstandard==0.23.0
